+++ src/Entities/AdjustmentCreditNote.php
@@ -1,0 +1,25 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * |------
+ * | ! Generated code !
+ * | Altering this code will result in changes being overwritten |
+ * |-------------------------------------------------------------|.
+ */
+
+namespace Paddle\SDK\Entities;
+
+class AdjustmentCreditNote
+{
+    private function __construct(
+        public string $url,
+    ) {
+    }
+
+    public static function from(array $data): self
+    {
+        return new self($data['url']);
+    }
+}