+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 Check our main [developer changelog](https://developer.paddle.com/?utm_source=dx&utm_medium=paddle-php-sdk) for information about changes to the Paddle Billing platform, the Paddle API, and other developer tools.
 
+## [1.3.0] - 2024-09-30
+
+### Added
+
+- Added `AdjustmentsClient::getCreditNotePDF`, see [related changelog](https://developer.paddle.com/changelog/2024/generate-adjustments-credit-notes)
+
 ## [1.2.0] - 2024-09-18
 
 ### Added